Refactor note image handling and modal close behavior
- Updated IndexingServiceTests to use GetByUserIdWithImagesAsync for fetching notes with images.
- Refactored CreateNoteModal and EditNoteModal to reset forms on close using a new handleClose function.
- Moved fileAttachmentsToNoteImages utility function to a new note-image-utils file for better organization and clarity.
- Removed redundant code from NoteImageAttachment component related to file attachment conversion.
